--- src/components/pagination/PaginationComponent/PaginationComponent.js
@@ -1,19 +1,0 @@
-import styles from "../Pagination.module.css"
-
-function PaginationComponent({ pages, setCurrentPage, currentPage }) {
-    return (
-        <div className={styles.pagination_div}>
-            {Array.from(Array(pages), (article, index) => {
-                return <button
-                style = {index === currentPage ? {backgroundColor: "lightgrey"} : null}
-                className={styles.pagination_button}
-                value={index}
-                onClick={
-                    (e) => setCurrentPage(Number(e.target.value))
-                }>{index + 1}</button>
-            })}
-        </div>
-    )
-}
-
-export default PaginationComponent;
